requirements.txt
streamlit==1.5.0
Pillow==9.0.0
tensorflow-cpu==2.7.0
